refactor(redis/balance-service): Move SettlementClient to Balance Service
This allows a store implementer to not have to worry about when to settle.
Convert trait was moved to settlement crate to avoid cyclic dependency
Fixes #116
