+++ crates/interledger-service-util/src/balance_service.rs
@@ -1,6 +1,8 @@
 use futures::Future;
+use interledger_ildcp::IldcpAccount;
 use interledger_packet::{Address, ErrorCode, Fulfill, Reject, RejectBuilder};
 use interledger_service::*;
+use interledger_settlement::{SettlementAccount, SettlementClient, SettlementStore};
 use log::{debug, error};
 use std::marker::PhantomData;
 use tokio_executor::spawn;
@@ -20,7 +22,7 @@ pub trait BalanceStore: AccountStore {
         &self,
         to_account: Self::Account,
         outgoing_amount: u64,
-    ) -> Box<dyn Future<Item = (), Error = ()> + Send>;
+    ) -> Box<dyn Future<Item = (i64, u64), Error = ()> + Send>;
 
     fn update_balances_for_reject(
         &self,
@@ -39,20 +41,22 @@ pub struct BalanceService<S, O, A> {
     ilp_address: Address,
     store: S,
     next: O,
+    settlement_client: SettlementClient,
     account_type: PhantomData<A>,
 }
 
 impl<S, O, A> BalanceService<S, O, A>
 where
-    S: BalanceStore,
+    S: BalanceStore<Account = A> + SettlementStore<Account = A>,
     O: OutgoingService<A>,
-    A: Account,
+    A: Account + SettlementAccount + IldcpAccount,
 {
     pub fn new(ilp_address: Address, store: S, next: O) -> Self {
         BalanceService {
             ilp_address,
             store,
             next,
+            settlement_client: SettlementClient::new(),
             account_type: PhantomData,
         }
     }
@@ -60,9 +64,9 @@ where
 
 impl<S, O, A> OutgoingService<A> for BalanceService<S, O, A>
 where
-    S: BalanceStore<Account = A> + Clone + Send + Sync + 'static,
+    S: BalanceStore<Account = A> + SettlementStore<Account = A> + Clone + Send + Sync + 'static,
     O: OutgoingService<A> + Send + Clone + 'static,
-    A: Account + 'static,
+    A: Account + IldcpAccount + SettlementAccount + 'static,
 {
     type Future = BoxedIlpFuture;
 
@@ -83,11 +87,15 @@ where
         let store_clone = store.clone();
         let from = request.from.clone();
         let from_clone = from.clone();
+        let from_id = from.id();
         let to = request.to.clone();
         let to_clone = to.clone();
+        let to_id = to.id();
         let incoming_amount = request.original_amount;
         let outgoing_amount = request.prepare.amount();
         let ilp_address = self.ilp_address.clone();
+        let settlement_client = self.settlement_client.clone();
+        let to_has_engine = to.settlement_engine_details().is_some();
 
         // Update the balance _before_ sending the settlement so that we don't accidentally send
         // multiple settlements for the same balance. While there will be a small moment of time (the delta
@@ -129,7 +137,25 @@ where
                             let fulfill_balance_update = store.update_balances_for_fulfill(
                                 to.clone(),
                                 outgoing_amount,
-                            ).map_err(move |_| error!("Error applying balance changes for fulfill from account: {} to account: {}. Incoming amount was: {}, outgoing amount was: {}", from.id(), to.id(), incoming_amount, outgoing_amount));
+                            )
+                            .map_err(move |_| error!("Error applying balance changes for fulfill from account: {} to account: {}. Incoming amount was: {}, outgoing amount was: {}", from_id, to_id, incoming_amount, outgoing_amount))
+                            .and_then(move |(_balance, amount_to_settle)| {
+                                if amount_to_settle > 0 && to_has_engine {
+                                    // Note that if this program crashes after changing the balance (in the PROCESS_FULFILL script)
+                                    // and the send_settlement fails but the program isn't alive to hear that, the balance will be incorrect.
+                                    // No other instance will know that it was trying to send an outgoing settlement. We could
+                                    // make this more robust by saving something to the DB about the outgoing settlement when we change the balance
+                                    // but then we would also need to prevent a situation where every connector instance is polling the
+                                    // settlement engine for the status of each
+                                    // outgoing settlement and putting unnecessary
+                                    // load on the settlement engine.
+                                    spawn(settlement_client
+                                        .send_settlement(to, amount_to_settle)
+                                        .or_else(move |_| store.refund_settlement(to_id, amount_to_settle)));
+                                }
+                                Ok(())
+                            });
+
                             spawn(fulfill_balance_update);
 
                             Ok(fulfill)
