+++ crates/interledger-service-util/src/exchange_rates_service.rs
@@ -1,8 +1,8 @@
-use crate::{Convert, ConvertDetails};
 use futures::{future::err, Future};
 use interledger_ildcp::IldcpAccount;
 use interledger_packet::{Address, ErrorCode, Fulfill, Reject, RejectBuilder};
 use interledger_service::*;
+use interledger_settlement::{Convert, ConvertDetails};
 use log::{error, trace};
 use std::marker::PhantomData;
