+++ crates/interledger-settlement/Cargo.toml
@@ -13,7 +13,6 @@ futures = "0.1.25"
 hyper = "0.12.29"
 interledger-packet = { path = "../interledger-packet", version = "0.2.1" }
 interledger-service = { path = "../interledger-service", version = "0.2.1" }
-interledger-service-util = { path = "../interledger-service-util", version = "0.2.1" }
 interledger-ildcp = { path = "../interledger-ildcp", version = "0.2.1" }
 log = "0.4.6"
 reqwest = "0.9.17"
