+++ crates/interledger-settlement/src/client.rs
@@ -1,10 +1,9 @@
-use super::SettlementAccount;
+use super::{Convert, ConvertDetails, SettlementAccount};
 use futures::{
     future::{err, Either},
     Future,
 };
 use interledger_ildcp::IldcpAccount;
-use interledger_service_util::{Convert, ConvertDetails};
 use log::{error, trace};
 use reqwest::r#async::Client;
 use serde_json::json;
